truffle develop, migrate ok

+++ contracts/Migrations.sol
@@ -1,22 +1,19 @@
 // SPDX-License-Identifier: MIT
-
 pragma solidity >=0.4.21 <0.7.0;
 
 contract Migrations {
     address public owner;
-    uint public last_completed_migration;
+    uint256 public last_completed_migration;
 
     modifier restricted() {
-        if(msg.sender == owner) _;
+        if (msg.sender == owner) _;
     }
 
-    constructor () public {
+    constructor() public {
         owner = msg.sender;
     }
-    
-    function setCompleted(uint completed) public restricted {
+
+    function setCompleted(uint256 completed) public restricted {
         last_completed_migration = completed;
     }
 }
-
-
